Cloud functions/google-metadata/requirements.txt
pandas
google-cloud-storage 
functions-framework	
google-auth	
google-auth-oauthlib 
pandas-gbq 
google-cloud-bigquery